+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.0
+
+**Breaking changes**:
+
+- Updated to use `http_client` 0.2.0, which has a slightly different
+  API to initialize the HTTP Client.
+
 ## 0.0.2
 
 - Enable url-form-encoding (for `POST` requests).